đông-xóa sản phẩm trong giỏ hàng-v2

+++ app/Http/Controllers/CartController.php
@@ -33,14 +33,42 @@ class CartController extends Controller
 
     public function deleteItemCart(Request $request, string $variant_id)
     {
+        $variant = ProductVariant::find($variant_id);
         $cart = session('cart')?session('cart'):null;
+        if(array_key_exists($variant_id,$cart->listProductVariants)==false){
+            return response()->json([
+                'sussess'=>false,
+                'message'=>'giỏ hàng chưa có sản phẩm này!'
+            ]);
+        }
         if($cart==null)
-            return 'giỏ hàng chưa có sản phẩm!';
+            return response()->json([
+                'sussess'=>false,
+                'giỏ hàng chưa có sản phẩm!'
+            ]);
         $cart->deleteItemCart($variant_id);
         if($cart->totalQuantity==0)
+        {
             $request->session()->forget('cart');
+            return response()->json([
+                'sussess'=>true,
+                'message'=> 'Đã xóa '.$variant->product->name.' ('.$variant->color.'/'.$variant->internal_memory.') khỏi giỏ hàng'
+            ]);
+        }
         else
             $request->session()->put('cart',$cart);
-        return 'Đã xóa '.$product->name.' ('.$variant->color.'/'.$variant->internal_memory.') khỏi giỏ hàng';
+        return response()->json([
+            'sussess'=>true,
+            'message'=> 'Đã xóa '.$variant->product->name.' ('.$variant->color.'/'.$variant->internal_memory.') khỏi giỏ hàng',
+            'cart'=>['totalQuantity'=>$cart->totalQuantity,'totalPrice'=>$cart->totalPrice]
+        ]);
+    }
+    public function deleteAllItem(Request $request){
+        if(session('cart')!=null)
+        {
+            $request->session()->forget('cart');
+            return 'Đã xóa tất cả sản phẩm trong giỏ hàng!';
+        }
+        return 'giỏ hàng chưa có sản phẩm!';
     }
 }
